scripts/requirements.txt
PyYAML
pydub
python-telegram-bot --upgrade
gTTS
requests